+++ Editor/Core/Data/R2EKEditorProjectSettingsProvider.cs
@@ -19,7 +19,8 @@ namespace RoR2.Editor
             var keywords = new[] { "RoR2EditorKit", "R2EK" };
             VisualElementTemplateDictionary.instance.DoSave();
             var settings = R2EKEditorProjectSettings.instance;
-            EditorSettingManager.PurgeOrphanedSettings(settings);
+            if (R2EKSettings.instance.purgeDefaultProviders)
+                EditorSettingManager.PurgeOrphanedSettings(settings);
             settings.hideFlags = HideFlags.DontSave | HideFlags.HideInHierarchy;
             settings.SaveSettings();
             return new R2EKEditorProjectSettingsProvider("Project/R2EK Editor Settings", SettingsScope.Project, keywords)
